pyproject.toml
[project]
name = "resolve-issue-agent"
version = "0.1.0"
description = "Agent that automates resolving GitHub issues by analyzing, implementing fixes, and creating pull requests"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "litellm>=1.68.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
